% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{find_delim}
\alias{find_delim}
\title{Find the delimiter of a file}
\usage{
find_delim(file_path = NULL)
}
\arguments{
\item{file_path}{character(1) path of the file.}
}
\value{
character(1), the delimiter found.
}
\description{
Find the delimiter of a file by reading the first 10 lines.
}
\author{
Damien Olivier

Rico Derks
}
