% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{histogram_plot}
\alias{histogram_plot}
\title{Histogram plot}
\usage{
histogram_plot(data = NULL)
}
\arguments{
\item{data}{list from output of prepare_histogram_data().}
}
\value{
ggplot2 object
}
\description{
Histogram plot.
}
\author{
Rico Derks
}
