% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{missing_plot}
\alias{missing_plot}
\title{Missing values plot}
\usage{
missing_plot(data = NULL, title = NULL)
}
\arguments{
\item{data}{data.frame from output of prepare_missing_data().}

\item{title}{character(1), title of the plot.}
}
\value{
ggplot2 object.
}
\description{
Missing values plot.
}
\author{
Rico Derks
}
