% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{pca_loadings_plot}
\alias{pca_loadings_plot}
\title{PCA loadings plot}
\usage{
pca_loadings_plot(data = NULL, xaxis = "PC1", yaxis = "PC2")
}
\arguments{
\item{data}{list from output of prepare_pca_data().}

\item{xaxis}{character(1), which PC to show on the x-axis.}

\item{yaxis}{character(1), which PC to show on the y-axis.}
}
\value{
ggplot2 object, the loadings plot.
}
\description{
PCA loadings plot.
}
\author{
Rico Derks
}
