% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{pca_scores_plot}
\alias{pca_scores_plot}
\title{PCA scores plot}
\usage{
pca_scores_plot(
  data = NULL,
  sampletype_col = NULL,
  batch_col = NULL,
  xaxis = "PC1",
  yaxis = "PC2"
)
}
\arguments{
\item{data}{list from output of prepare_pca_data().}

\item{sampletype_col}{character(1), name of the sample type column.}

\item{batch_col}{character(1), name of the batch column.}

\item{xaxis}{character(1), which PC to show on the x-axis.}

\item{yaxis}{character(1), which PC to show on the y-axis.}
}
\value{
ggplot2 object, the scores plot with density plots around it.
}
\description{
PCA scores plot.
}
\author{
Rico Derks
}
