% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{prepare_missing_data}
\alias{prepare_missing_data}
\title{Prepare data for a missing values plot}
\usage{
prepare_missing_data(
  data = NULL,
  meta_data = NULL,
  sampleid_raw_col = NULL,
  sampleid_meta_col = NULL,
  sample_ids = NULL
)
}
\arguments{
\item{data}{data.frame in wide format.}

\item{meta_data}{data.frame with the meta data.}

\item{sampleid_raw_col}{character(1), name of the sample id column in the raw data.}

\item{sampleid_meta_col}{character(1), name of the sample id column in the meta data.}

\item{sample_ids}{character() vector with the names of the samples id's which should be selected for the plot.}
}
\value{
data.frame
}
\description{
Prepare data for a missing values plot.
}
\author{
Rico Derks
}
