% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{prepare_pca_data}
\alias{prepare_pca_data}
\title{Do PCA and prepare the data for plotting}
\usage{
prepare_pca_data(
  data = NULL,
  meta_data = NULL,
  sampleid_raw_col = NULL,
  sampleid_meta_col = NULL,
  id_samples = NULL,
  id_qcpool = NULL
)
}
\arguments{
\item{data}{data.frame in wide format.}

\item{meta_data}{data.frame with the meta data.}

\item{sampleid_raw_col}{character(1), name of the sample id column in the raw data.}

\item{sampleid_meta_col}{character(1), name of the sample id column in the meta data.}

\item{id_samples}{character() vector with the names of the sample samples id's.}

\item{id_qcpool}{character() vector with the names of the pooled sample id's.}
}
\value{
list with the PCA model, summary of fit, scores and loadings.
}
\description{
Do PCA and prepare the data for plotting.
}
\author{
Rico Derks
}
