% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{prepare_trend_data}
\alias{prepare_trend_data}
\title{Prepare data for trend plot}
\usage{
prepare_trend_data(
  data = NULL,
  meta_data = NULL,
  sampleid_raw_col = NULL,
  sampleid_meta_col = NULL,
  order_col = NULL,
  batch_col = NULL,
  id_qcpool = NULL
)
}
\arguments{
\item{data}{data.frame in wide format.}

\item{meta_data}{data.frame with the meta data.}

\item{sampleid_raw_col}{character(1), name of the sample id column in the raw data.}

\item{sampleid_meta_col}{character(1), name of the sample id column in the meta data.}

\item{order_col}{character(1), name of the acquistion order column.}

\item{batch_col}{character(1), name of the batch column.}

\item{id_qcpool}{character() vector with the names of the pooled sample id's.}
}
\value{
data.frame in long format
}
\description{
Prepare data (wide format) for creating a trend plot (long format).
}
\details{
The reference is the first QCpool (pooled sample) measured in the all data set.
}
\author{
Rico Derks
}
