% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_data.R
\name{read_data}
\alias{read_data}
\title{Read the data file}
\usage{
read_data(file_path = NULL, sep = NULL, first_column_as_index = FALSE)
}
\arguments{
\item{file_path}{character(1) containing the file path to the imported data
file.}

\item{sep}{character(1) for several file formats a separator can be specified.}

\item{first_column_as_index}{logical(1) should the first column be used as
index.}
}
\value{
data.frame
}
\description{
Read the data file. It can be in different formats
}
\details{
data
}
\author{
Damien Olivier

Rico Derks
}
