% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{rle_plot}
\alias{rle_plot}
\title{Relative log expression plot}
\usage{
rle_plot(data = NULL, sampleid_raw_col = NULL, batch_col = NULL)
}
\arguments{
\item{data}{data.frame from output of prepare_rle_data().}

\item{sampleid_raw_col}{character(1), name of the sample id column in the raw data.}

\item{batch_col}{character(1), name of the batch column in the meta data.}
}
\value{
ggplot2 object.
}
\description{
Relative log expression plot.
}
\author{
Rico Derks
}
