% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_visualization.R
\name{trend_plot}
\alias{trend_plot}
\title{Create trend plot}
\usage{
trend_plot(
  data = NULL,
  sampleid_raw_col = NULL,
  batch_col = NULL,
  yaxis = c("log2fc", "log2fc_batch")
)
}
\arguments{
\item{data}{data.frame in long format created by prepare_trend_data().}

\item{sampleid_raw_col}{character(1), name of the sample id column in the raw data.}

\item{batch_col}{character(1), name of the batch column.}

\item{yaxis}{character(1), one of "log2fc" or "log2fc_batch".}
}
\value{
A plotly object
}
\description{
Create trend plot with plotly
}
\details{
visualization
}
\author{
Rico Derks
}
